1/3/25

+++ HW1/Quadratic.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Quadratic {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
-        double a = input.nextInt();
+        double a = input.nextDouble();
 
         System.out.println(Math.sqrt(a));
     }
